projeto-django/requirements.txt
Django==2.2.4
mysqlclient==1.4.2
pytz==2019.2
sqlparse==0.3.0
mysql-connector-python-rf==2.2.2
PyJWT==1.7.1 
